Move getNodesForAddress() to editor module
